backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
python-dotenv>=1.0.1
openai>=1.40.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
faiss-cpu>=1.8.0
